refactor(products): share duplicate-safe feature keys
