+++ src/components/products.tsx
@@ -21,6 +21,7 @@ import {
   type ProductSortOption,
   type ViewMode,
 } from "@/components/ui/filter-controls";
+import { buildFeatureItems } from "@/lib/feature-items";
 import { resolveProductImagePath } from "@/lib/image-utils";
 import type { ExtendedProduct } from "@/lib/types";
 
@@ -52,6 +53,9 @@ function ProductCard({ product }: ProductCardProps) {
 
   const heroImage = product.screenshots[0];
   const imagePath = resolveProductImagePath(heroImage, product.slug);
+  const previewFeatureItems = buildFeatureItems(
+    product.features?.slice(0, 4) ?? [],
+  );
 
   useEffect(() => {
     const ctx = gsap.context(() => {
@@ -183,9 +187,9 @@ function ProductCard({ product }: ProductCardProps) {
 
             {/* Features Preview */}
             <div className="grid grid-cols-1 sm:grid-cols-2 gap-5 mb-10">
-              {product.features?.slice(0, 4).map((feature, idx) => (
+              {previewFeatureItems.map(({ feature, key }, idx) => (
                 <motion.div
-                  key={`${feature}-${idx}`}
+                  key={key}
                   initial={{ opacity: 0, x: -10 }}
                   whileInView={{ opacity: 1, x: 0 }}
                   transition={{ delay: 0.1 * idx }}
