+++ src/components/products/product-features.tsx
@@ -2,28 +2,10 @@
 
 import { CheckCircle } from "@phosphor-icons/react";
 import { motion } from "framer-motion";
+import { buildFeatureItems } from "@/lib/feature-items";
 
 interface ProductFeaturesProps {
   features: string[];
-}
-
-interface FeatureItem {
-  feature: string;
-  key: string;
-}
-
-export function buildFeatureItems(features: string[]): FeatureItem[] {
-  const occurrenceCounts = new Map<string, number>();
-
-  return features.map((feature) => {
-    const occurrence = occurrenceCounts.get(feature) ?? 0;
-    occurrenceCounts.set(feature, occurrence + 1);
-
-    return {
-      feature,
-      key: `${feature.length}:${feature}:${occurrence}`,
-    };
-  });
 }
 
 /**
